Update Navbar.jsx and index.css
Changed most elements of the Navbar.jsx to organize and be functional with the index.css. Also added a function called nav() to open and close the navbar when in mobile version.
Also, made changes to the index.css to make a functional mobile version navbar, as well as a hover dropdown for both desktop and mobile versions.